requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LangChain
langchain==0.0.340
langchain-openai==0.0.2
openai==1.3.8

# Base de données
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Services
redis==5.0.1
celery==5.3.4

# Audio/Vidéo
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11
opencv-python==4.8.1.78
pillow==10.1.0

# Utilitaires
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4